chore: codebase-wide quality, readability, and docs accuracy sweep
Three coordinated passes across the SDK + dashboard:

Quality (round 1-3 + 8 parallel agents):
- 25+ real bugs fixed: F821 type annotations, B039 mutable ContextVar
  default, B012 return-in-finally exception suppression, F841 broken
  export_spans return path, F402 shadowed import, RUF012 mutable class
  defaults in 3 trace instrumentors, RUF013 implicit Optional, RUF006
  asyncio dangling task, O(n^2) Gini in annotation_delegation, unbounded
  _parent_id_map LRU cap, _safe_value cyclic-input infinite loop,
  numeric_rmse_metric returning abs error instead of squared error
- Misc perf nits across cli/commands, calibration, metrics, trace
- Deleted DC-006 orphan example_agents/.../app.py + llm_callers.py
- 11 stale CODE_AUDIT.md findings marked [stale] / [resolved]

Readability:
- ruff format applied to entire codebase (502 files reformatted)
- pyproject.toml ignore comment updated to record the format pass

Docs accuracy:
- README, DESIGN, CONTRIBUTING: dashboard/report rename corrected,
  stale FastAPI-aspirational note removed, file tree fixed
- sdk/README.md, sdk/CHANGELOG.md: command rename + dead doc ref
- docs/clis/: report.md created (missing after rename), dashboard.md
  output default fixed, README index updated
- docs/dev/research/ape-investigation.md: parameter table now matches
  APEConfig defaults
- docs/technical-manual.md: removed reference to deleted llm_callers

Memory notes:
- feedback_code_audit_drift.md - verify audit findings against source
- project_codebase_shape.md - 188/189 top-level SDK files are
  test-paired scaffolding orphans; live surface is ~200 files

+++ dashboard/evalyn_dashboard/api/cli.py
@@ -51,9 +51,7 @@ def _coerce_number(value: Any) -> float | int | None:
     return None
 
 
-def _collect_validation_errors(
-    schema: CliSchema, args: dict[str, Any]
-) -> list[str]:
+def _collect_validation_errors(schema: CliSchema, args: dict[str, Any]) -> list[str]:
     """Non-raising counterpart of ``_validate_args``. Returns a list of
     human-readable error strings; empty list means valid.
 
@@ -151,9 +149,7 @@ def _param_to_argv(param: ParamSchema, value: Any) -> list[str]:
         if param.options and value not in param.options:
             raise HTTPException(
                 status_code=400,
-                detail=(
-                    f"{param.name}: {value!r} not in {param.options}"
-                ),
+                detail=(f"{param.name}: {value!r} not in {param.options}"),
             )
         return [flag, str(value)]
 
@@ -222,6 +218,7 @@ async def get_catalog(
     if q is not None:
         needle = q.strip().lower()
         if needle:
+
             def _matches(s: CliSchema) -> bool:
                 # description and blurb are documented optional fields
                 # on CliSchema; getattr keeps us robust to dataclasses
@@ -330,9 +327,7 @@ async def get_command_history(
             args = row.get("args") if isinstance(row.get("args"), dict) else {}
             duration_s = row.get("duration_s")
             duration_ms = (
-                int(round(float(duration_s) * 1000.0))
-                if isinstance(duration_s, (int, float))
-                else 0
+                round(float(duration_s) * 1000.0) if isinstance(duration_s, (int, float)) else 0
             )
             status = _persistence_status_to_run_status(
                 str(row.get("status") or ""), row.get("exit_code")
